refactor: move existing fs implementation to internal _fs directory (#4381)
