+++ std/node/_fs/_fs_access.ts
@@ -1,0 +1,23 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+
+import { CallbackWithError } from "./_fs_common.ts";
+import { notImplemented } from "../_utils.ts";
+
+/** Revist once https://github.com/denoland/deno/issues/4017 lands */
+
+//TODO - 'path' can also be a Buffer or URL.  Neither of these polyfills
+//is available yet.  See https://github.com/denoland/deno/issues/3403
+export function access(
+  path: string, // eslint-disable-line @typescript-eslint/no-unused-vars
+  modeOrCallback: number | Function, // eslint-disable-line @typescript-eslint/no-unused-vars
+  callback?: CallbackWithError // eslint-disable-line @typescript-eslint/no-unused-vars
+): void {
+  notImplemented("Not yet available");
+}
+
+//TODO - 'path' can also be a Buffer or URL.  Neither of these polyfills
+//is available yet.  See https://github.com/denoland/deno/issues/3403
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function accessSync(path: string, mode?: number): void {
+  notImplemented("Not yet available");
+}
