+++ std/node/_fs/testdata/hello.txt
